requirements.txt
autopep8==1.7.0
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
